requirements.txt
pyinstaller==6.18.0
PySide6==6.10.1
PyYAML==6.0.3